[fix] tensor label type

+++ src/storage/datasetrepo.py
@@ -79,9 +79,9 @@ class GTSRBRepository(BaseRepository):
         x_train = x_train.reshape((x_train.shape[0], x_train.shape[3], x_train.shape[1], x_train.shape[2]))
         x_test = x_test.reshape((x_test.shape[0], x_test.shape[3], x_test.shape[1], x_test.shape[2]))
         train_dataset = TransformableDataset(
-             torch.tensor(x_train, dtype=torch.float), torch.tensor(y_train, dtype=torch.uint8)
+             torch.tensor(x_train, dtype=torch.float), torch.tensor(y_train, dtype=torch.long)
         )
         test_dataset = TransformableDataset(
-            torch.tensor(x_test, dtype=torch.float), torch.tensor(y_test, dtype=torch.uint8)
+            torch.tensor(x_test, dtype=torch.float), torch.tensor(y_test, dtype=torch.long)
         )
         return train_dataset, test_dataset
